requirements.txt
# Base-------------------------------------
albumentations>=1.1.0
ipython
jupyter
matplotlib
opencv-python>=4.1.1.26
Pillow
PyYAML
scikit-image
scikit-learn
scipy
torch>=1.12.0
torchvision>=0.13.0
numpy
protobuf<=3.20.1
pandas
tqdm
deep-sort-realtime
gradio==3.50.2
albumentations==1.1.0